+++ scripts/ShelfControl/messages/npcOwned.lua
@@ -30,7 +30,7 @@ local unlockableRules = {
     },
     {
         cond = function(ctx)
-            local ench = ctx.player.type.skills.attributes.enchant(ctx.player)
+            local ench = ctx.player.type.stats.skills.enchant(ctx.player)
             return ench.modified >= HIGH_ENCH
         end,
         key = "highEnch"
@@ -40,7 +40,7 @@ local unlockableRules = {
 local specificRules = {
     {
         cond = function(ctx)
-            local class = GetRecord(ctx.owner).class
+            local class = GetRecord(ctx.owner.self).class
             return MagicClasses[class] == true
         end,
         key = "ownerIsOfMagicClass",
@@ -53,21 +53,16 @@ local specificRules = {
 
 local function collectGenericMessages(subgroups, ctx)
     local prefix = msgSrc .. "generic"
-    local name = GetRecord(ctx.owner.self).name
-    return CollectAllMessagesByPrefix(prefix, l10n, {npc_name = name})
-end
-
-local function collectEggMessages(subgroups, ctx)
-    local prefix = msgSrc .. "easterEgg"
-    return CollectAllMessagesByPrefix(prefix, l10n)
+    return CollectAllMessagesByPrefix(
+        prefix, l10n, AdditionalMsgCtx(GetRecord(ctx.owner.self)))
 end
 
 local function collectUnlockableMessages(subgroups, ctx)
     local msgs = {}
-    local name = GetRecord(ctx.owner.self).name
     for group, _ in pairs(subgroups) do
         local prefix = msgSrc .. "unlockable_" .. group
-        local collectedMsgs = CollectAllMessagesByPrefix(prefix, l10n, {npc_name = name})
+        local collectedMsgs = CollectAllMessagesByPrefix(
+            prefix, l10n, AdditionalMsgCtx(GetRecord(ctx.owner.self)))
         AppendArray(msgs, collectedMsgs)
     end
     return msgs
@@ -75,10 +70,10 @@ end
 
 local function collectSpecificMessages(subgroups, ctx)
     local msgs = {}
-    local name = GetRecord(ctx.owner.self).name
     for group, _ in pairs(subgroups) do
         local prefix = msgSrc .. "specific_" .. group
-        local collectedMsgs = CollectAllMessagesByPrefix(prefix, l10n, {npc_name = name})
+        local collectedMsgs = CollectAllMessagesByPrefix(
+            prefix, l10n, AdditionalMsgCtx(GetRecord(ctx.owner.self)))
         AppendArray(msgs, collectedMsgs)
     end
     return msgs
@@ -86,10 +81,14 @@ end
 
 local msgCollectors = {
     generic = collectGenericMessages,
-    easterEgg = collectEggMessages,
+    easterEgg = CollectEggMessages,
     unlockable = collectUnlockableMessages,
     specific = collectSpecificMessages,
 }
+
+-- +----------+
+-- | The core |
+-- +----------+
 
 function PickNPCOwnedMessage(ctx)
     local msgGroups = {