+++ scripts/ShelfControl/settingsGlobal.lua
@@ -92,8 +92,9 @@ I.Settings.registerGroup {
         {
             key = 'ignoreScrolls',
             name = 'ignoreScrolls_name',
+            description = 'ignoreScrolls_description',
             renderer = 'checkbox',
-            default = false,
+            default = true,
         },
         {
             key = 'enableDebug',